Add Datahub access token to the environment
Data hub APIs protected requiring user to authenticate first.
For development purpose we can use the generated access token
from Django admin and use it to send requests to the API.

We should not commit the real token to the repository, it should
be used mainly for development purposes only.