Add helper function to do a Company search using Data Hub API
When the user is editing an enquiry we will do a search on
Data Hub and display possible matches with that company name.
This is to help the user match with the correct company (if
it exists on Data Hub). Add helper function to perform Company
search using Data Hub.

When the user is editing a company we would like to know whether
the company already exists in the Data hub, so perform a company
search and return results if any.