Add separate settings file for Tests
We have to override few settings during tests eg page_size for pagination.
Overriding in this case doesn't work because the view is initialized
at the time of module load and it doesn't pickup overrided settings.
It is convenient to have a separate settings file for tests and specify
test settings in this file.