Fetch and process enquiries periodically
Periodically fetch enquiries from AS using a celery task and create
instances in our database. We track the last processed enquiry so that
we can fetch new data after that enquiry.

We process individually so as to record the failed one. This data can
be provided to dev/service teams to understand the reasons for failure
and populate them in the system after correction.