Fix processing of CSV upload with different encodings
Depending on the user operating system csv could be in different
encodings (utf-8, windows-1252 etc). Since we process the file on
PaaS we cannot determine User OS hence assume utf-8 by default and
if there is an issue switch to windows-1252.

It is to be noted that line endings will also be different in each OS.
Eg Windows and Mac uses '\r\n' where as linux uses '\n'

Fixed failing tests.