Refactor and add additional error checks
Simplified checking for a valid csv file type and additional error checks
such as skipping the file if it is empty or contains only header row.

This is not necessarily an error as we can simply ignore and won't be
able to import anything but flag it as error so user can take action.