Fix pagination in combination with filters
Both pagination and filters use query parameters to provide variables
to backend. They both individually work ok but the parameters gets
overwritten if used in combination. Eg apply some filters and select
a particular page from the list of results. In this case once a page
is selected then existing query parameters gets overwritten with the
selected page and because of this filters also gets cleared.

The fix is to append the variables instead of overwriting them which
is implemented using custom template tags.