Fix company name filter not showing full search term
If the company name search is performed with multiple words separated
by space then upon fetching the results only the first word is shown
in the search field. The results themselves are correct except that
only first word is shown in the search field.

This is just due to not enclosing input field value in double quotes.