CPS-378: catch http errors correctly
The consent service was returning 404s for missing consent data, which
were not being caught correctly and causing the page to return a 500. It
is possible for people not to exist on the consent service if the
enquiries were added manually (rather than through the Activity Stream)
so we want to handle 404s and return False.

The code coverage tool is alerting that this code is not covered by
tests, but we have written unit tests to cover it. The e2e specs do not
currently have the setting `FEATURE_ENFORCE_CONSENT_SERVICE` on, so no
consent integration has ever been tested, and it's outside of the scope
of this bug fix.