Fix 403 error by using the existing, custom, HawkAuth class
The addition of the payload to the GET request was causing the API to return a 403 error. However, removing the payload caused an exception from the `requests_hawk` package, requiring a payload when `always_hash_content=True`. Instead, we were able to resolve the issue by using the custom HawkAuth class, which is also used elsewhere in the EMT application.